Added support for the second generation - Shelly "plus"
Since the API for the plus devices is different, I've split up the functionality so that the functionality for polling the first generation is in one file, and the second gen in another.

The structure of these files is largely the same, and you still access both of them via the "metrics" entry point.

Key differences:
  - The new generation uses digest authentication rather than basic, and has a hardcoded username
  - The URL has changed, they now use an RPC style API
  - The format of the json you get back is also different (but standardised across all the new generation

I've also added support for relays with multiple switches (i.e. PM2). I'm not able to test it extensively (yet), as all the relays I have are only of the single variety, but I'm fairly confident it should work, as it was only a minor tweak to the script